requirements.txt
Flask==2.0.3
requests==2.26.0
pytest==6.2.5
ipykernel
langchain
langchain-openai
python-dotenv
scikit-learn
azure-ai-projects
azure-identity
openai
streamlit